Refactor to split core/gui classes
